+++ source/packages/@aws-accelerator/accelerator/lib/asea-resources/shared-security-groups.ts
@@ -2,10 +2,13 @@ import { AseaResourceType, VpcConfig, VpcTemplatesConfig } from '@aws-accelerato
 import { SsmResourceType } from '@aws-accelerator/utils/lib/ssm-parameter-path';
 import { AseaResource, AseaResourceProps } from './resource';
 import { ImportAseaResourcesStack, LogLevel } from '../stacks/import-asea-resources-stack';
+import { AcceleratorStage } from '../accelerator-stage';
 import { pascalCase } from 'pascal-case';
 
 const enum RESOURCE_TYPE {
   SECURITY_GROUP = 'AWS::EC2::SecurityGroup',
+  SECURITY_GROUP_EGRESS = 'AWS::EC2::SecurityGroupEgress',
+  SECURITY_GROUP_INGRESS = 'AWS::EC2::SecurityGroupIngress',
 }
 
 const ASEA_PHASE_NUMBER = '2';
@@ -22,6 +25,9 @@ export class SharedSecurityGroups extends AseaResource {
     }
     const vpcsInScope = this.scope.sharedVpcs;
     this.updateSecurityGroups(vpcsInScope);
+    if (this.props.stage === AcceleratorStage.POST_IMPORT_ASEA_RESOURCES) {
+      this.deleteSharedSecurityGroups(vpcsInScope);
+    }
   }
   private updateSecurityGroups(vpcItems: (VpcConfig | VpcTemplatesConfig)[]) {
     if (vpcItems.length === 0) {
@@ -64,5 +70,125 @@ export class SharedSecurityGroups extends AseaResource {
       }
     }
     return;
+  }
+
+  private deleteSharedSecurityGroups(vpcItems: (VpcConfig | VpcTemplatesConfig)[]) {
+    if (vpcItems.length === 0) {
+      return;
+    }
+    for (const vpcItem of vpcItems) {
+      const vpcStackInfo = this.getSecurityGroupResourceByVpc(vpcItem.name);
+      if (!vpcStackInfo) {
+        continue;
+      }
+      const existingSecurityGroups = vpcStackInfo.nestedStackResources.getResourcesByType(RESOURCE_TYPE.SECURITY_GROUP);
+
+      const existingSecurityGroupIngressRules = vpcStackInfo.nestedStackResources.getResourcesByType(
+        RESOURCE_TYPE.SECURITY_GROUP_INGRESS,
+      );
+      const existingSecurityGroupEgressRules = vpcStackInfo.nestedStackResources.getResourcesByType(
+        RESOURCE_TYPE.SECURITY_GROUP_EGRESS,
+      );
+
+      for (const existingSecurityGroup of existingSecurityGroups) {
+        const existingSecurityGroupName = existingSecurityGroup.resourceMetadata['Properties']['GroupName'];
+        const securityGroupConfig = vpcItem.securityGroups?.find(
+          (securityGroupItem: { name: string }) => securityGroupItem.name === existingSecurityGroupName,
+        );
+        // if the security group is still configured skip
+        if (securityGroupConfig) continue;
+
+        // lookup security group by tag name
+        // if unable to locate, warn and then skip
+        const securityGroupResource = vpcStackInfo.nestedStackResources.getResourceByTag(existingSecurityGroupName);
+        if (!securityGroupResource) {
+          this.scope.addLogs(
+            LogLevel.WARN,
+            `Could not locate shared security group resouce by tag name : ${existingSecurityGroupName}, for vpc ${vpcItem.name}`,
+          );
+          continue;
+        }
+
+        // remove security group from template
+        this.scope.addLogs(LogLevel.WARN, `Deleting Shared Security Group: ${existingSecurityGroup.logicalResourceId}`);
+        this.scope.addDeleteFlagForNestedResource(
+          vpcStackInfo.nestedStackResources.getStackKey(),
+          securityGroupResource.logicalResourceId,
+        );
+
+        const ssmResource = vpcStackInfo.nestedStackResources.getSSMParameterByName(
+          this.scope.getSsmPath(SsmResourceType.SECURITY_GROUP, [
+            vpcItem.name,
+            existingSecurityGroup.resourceMetadata['Properties'].GroupName,
+          ]),
+        );
+        if (ssmResource) {
+          this.scope.addLogs(LogLevel.WARN, `Deleting SSM Parameter: ${ssmResource.logicalResourceId}`);
+          this.scope.addDeleteFlagForNestedResource(
+            vpcStackInfo.nestedStackResources.getStackKey(),
+            ssmResource.logicalResourceId,
+          );
+        }
+
+        for (const ingressRule of existingSecurityGroupIngressRules) {
+          try {
+            if (ingressRule.resourceMetadata['Properties'].GroupId['Ref'] === existingSecurityGroup.logicalResourceId) {
+              this.scope.addLogs(LogLevel.WARN, `Deleting Ingress Rule: ${ingressRule.logicalResourceId}`);
+              this.scope.addDeleteFlagForNestedResource(
+                vpcStackInfo.nestedStackResources.getStackKey(),
+                ingressRule.logicalResourceId,
+              );
+            }
+          } catch (error) {
+            // continue the ref may not exits
+          }
+
+          try {
+            if (
+              ingressRule.resourceMetadata['Properties'].SourceSecurityGroupId['Ref'] ===
+              existingSecurityGroup.logicalResourceId
+            ) {
+              this.scope.addLogs(LogLevel.WARN, `Deleting Ingress Rule: ${ingressRule.logicalResourceId}`);
+              this.scope.addDeleteFlagForNestedResource(
+                vpcStackInfo.nestedStackResources.getStackKey(),
+                ingressRule.logicalResourceId,
+              );
+            }
+          } catch (error) {
+            // the ref may not exist
+          }
+
+          for (const egressRule of existingSecurityGroupEgressRules) {
+            try {
+              if (
+                egressRule.resourceMetadata['Properties'].GroupId['Ref'] === existingSecurityGroup.logicalResourceId
+              ) {
+                this.scope.addLogs(LogLevel.WARN, `Deleting Egress Rule: ${egressRule.logicalResourceId}`);
+                this.scope.addDeleteFlagForNestedResource(
+                  vpcStackInfo.nestedStackResources.getStackKey(),
+                  egressRule.logicalResourceId,
+                );
+              }
+            } catch (error) {
+              // continue the ref may not exist
+            }
+            try {
+              if (
+                egressRule.resourceMetadata['Properties'].DestinationSecurityGroupId['Ref'] ===
+                existingSecurityGroup.logicalResourceId
+              ) {
+                this.scope.addLogs(LogLevel.WARN, `Deleting Egress Rule: ${egressRule.logicalResourceId}`);
+                this.scope.addDeleteFlagForNestedResource(
+                  vpcStackInfo.nestedStackResources.getStackKey(),
+                  egressRule.logicalResourceId,
+                );
+              }
+            } catch (error) {
+              // continue the ref may not exist
+            }
+          }
+        }
+      }
+    }
   }
 }